First iteration of install, retrieve package by name, and extract data
I believe that extract data might be useful to facilitate transfert of information to the front-end later on. Install works but still need to add callback, I am planning to provide ProgressCallback into the constructor, so it could be easier to give you a callback ( just a thought ). Able to find a snap using the name