Update PackageKitService with search, install (WIP), remove (WIP) package
It has most of the implementation to use install and remove but still need to add a few tests to make sure everything works as expected. Implements the search method for the packagekitglib. Use the earlier documentation of how a package should be formatted in this class and add it as a class pydoc since it is the same rule for installing or removing a package. Move the new methods in a more appropriate order inside the class